feat: features 레이어 예제 추가 (#3)
