+++ src/features/get-shared-post/index.ts
@@ -1,0 +1,1 @@
+export * from './get-shared-post.api';
