+++ src/features/get-shared-posts/get-shared-posts.api.ts
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getSharedPosts } from '@/entities/shared-post/api';
+import { sharedPostQueryKey } from '@/shared';
+
+export const useSharedPostsQuery = () =>
+  useQuery({
+    queryKey: [sharedPostQueryKey],
+    queryFn: async () => await getSharedPosts(),
+  });
