+++ src/features/get-shared-posts/index.ts
@@ -1,0 +1,1 @@
+export * from './get-shared-posts.api';
