--- src/features/shared-post/index.ts
@@ -1,1 +1,0 @@
-export * from './shared-post.api';
