+++ src/features/update-shared-post/index.ts
@@ -1,0 +1,1 @@
+export * from './update-shared-post.api';
