+++ src/shared/query-keys.ts
@@ -1,0 +1,1 @@
+export const sharedPostQueryKey = '/shared/posts';
